Removed the conditional return statements on  method. Reason is that, we should be able to force it to scroll to the bottom. I'm not even sure why I had it in there in the first place.
Other changes:
Changed @isDead to @stopped because the former's ambiguity.

Added method chaining for the @addEvents method and replaced 'bind' with 'on'.
Similar thing done to @removeEvents

Cleaned up code and added comments on the @reset method.

It may have added a few weight in terms of file size, but I prefer
having a comprehensive code that's easy to improve and update.

Added explanation to the formula for the scrolling mechanism in the
@scroll method.

Added makefile for development preview.

On a personal note: Seriously, good job guys. I've seen the changes and just wanna say
thank you for contributing and helping the plugin work better. Special shoutout to @kristerkari
for the continuous updates.