Add conditional return statements for event handlers. As to prevent producing javascript errors, I have added a condition where it ignores any lines of code that uses the event callback data (see events.scroll and events.wheel).
I have also enabled the settings to be applied even after the first
definition of the scrollbar data of a jquery object.
(see line 265 of coffeescripts/jquery.nanoscroller.coffee).
This is done so that developers can later enable settings on a
later phase. Example scenario would be a global settings for all
the nanoScrollers present in the page while still maintaining a
unique settings for each of them if needed.